feat: update data types from Integer to long

+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/dataanalysis/DataAnalysisMeasures.java
@@ -48,7 +48,7 @@ public class DataAnalysisMeasures
     // Constructor
     // -------------------------------------------------------------------------
 
-    public DataAnalysisMeasures( int orgUnitId, int categoryOptionComboId,
+    public DataAnalysisMeasures( long orgUnitId, long categoryOptionComboId,
         double average, double standardDeviation )
     {
         this.orgUnitId = orgUnitId;
